requirements.txt
ctransformers==0.2.27
sentence-transformers==3.2.1
pinecone[grpc]
langchain
flask
pypdf
python-dotenv
langchain_pinecone
langchain_community
langchain_experimental
-e .